% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{area_picker}
\alias{area_picker}
\title{area_picker}
\usage{
area_picker(df)
}
\arguments{
\item{df}{}
}
\description{
area_picker
}
\examples{
df <- echem_read(system.file('extdata/cv/cv_example.txt', package = 'osc'))
df <- area_picker(df)
}
