% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{coverage}
\alias{coverage}
\title{Calculate coverage based on peak integration}
\usage{
coverage(df, diameter = NULL, area = NULL, electrons = 1)
}
\description{
Calculate coverage based on peak integration
}
\examples{
None yet
}
