% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cv_read}
\alias{cv_read}
\title{Load cyclic voltammogram}
\usage{
cv_read(file, skip, col_names = c("pot", "cur"), ...)
}
\arguments{
\item{skip}{Number of lines to skip (where does metadata end and the data start?)}

\item{path}{Path to the cv file (txt)}
}
\description{
Reads *.txt files from the CHI potentiostats. Automatically detects sweep- and cv-number, so these can later be used for colors in plots etc.
}
\examples{
file1 <- "C:/Users/emilbp/PhD/Projects/High-Shear Exfoliation/Data/20150602 CV of benzophenone/CV2 2mM benzophenone.txt"
df <- cv_read(file1, skip = 41)
}
\keyword{cyclic}
\keyword{electrochemistry}
\keyword{voltammetry,}
