% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrochemistry.R
\name{echem_read}
\alias{echem_read}
\title{Read electrochemistry data}
\usage{
echem_read(file, type = NA)
}
\arguments{
\item{file}{Path to the *.txt file from the CHI potentiostat}
}
\description{
Reads *.txt files from the CHI potentiostats.
Current supports Cyclic Voltammetry and Bulk Electrolysis with Coulometry
}
\examples{
file <- system.file('extdata/cv/cv_example.txt', package = 'osc')
data <- read_elec(elec)
plot(data)

file <- system.file('extdata/cv/Electrolysis.txt', package = 'osc')
data <- read_elec(file)
plot(data)

}
