% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{electrolysis_read}
\alias{electrolysis_read}
\title{Load electrolysis
Reads *.txt files from the CHI potentiostats containing electrolysis data.}
\usage{
electrolysis_read(file, skip, col_names = c("time", "charge", "current"))
}
\arguments{
\item{skip}{Number of lines to skip (where does metadata end and the data start?)}

\item{col_names}{A vector containing the names to be used for the two loaded columms (Default: c('time', 'charge', 'current'))}

\item{path}{Path to the CV file (txt)}
}
\description{
Load electrolysis
Reads *.txt files from the CHI potentiostats containing electrolysis data.
}
\examples{
file <- system.file('extdata/cv/cv_example.txt', package = 'osc')
df <- cv_read(file, skip = 41)

file <- system.file('extdata/cv/5cv_example.txt', package = 'osc')
df <- cv_read(file, skip = 70)
plot(df)
}
\seealso{
Other cyclic voltammetry, electrochemistry: \code{\link{cv_read}}
}
\keyword{cyclic}
\keyword{electrochemistry}
\keyword{voltammetry,}
