\name{exampledata}
\alias{exampledata}
\docType{data}
\title{
Example data for the clustering algorithm. 
}
\description{
This is test data for the package.
}
\usage{data(exampledata)}
\format{
  A data frame with 235 observations on the following 4 variables.
  \describe{
    \item{\code{x}}{a numeric vector of x-coordinates}
    \item{\code{y}}{a numeric vector of y-coordinates}
    \item{\code{z}}{a numeric vector of population data}
    \item{\code{cluster}}{a numeric vector of clusters}
  }
}
\examples{
data(exampledata)
str(exampledata)
}
\keyword{datasets}
