% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrochemistry.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata from the header of a datafile}
\usage{
extract_metadata(header, start, pattern = "[-+]?\\\\d+\\\\.?\\\\d*e?-?\\\\d*",
  numeric = TRUE)
}
\arguments{
\item{header}{A vector of header lines in a datafile}

\item{start}{A string with the beginning of the line of interest}

\item{pattern}{Regular expression detailing how to extract the value. The default will attempt to match any number.}

\item{numeric}{Should the function return a numeric (TRUE) or character (FALSE)}
}
\description{
Extract metadata from the header of a datafile
}
