% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrochemistry.R
\name{find_data}
\alias{find_data}
\title{Find the data startposition}
\usage{
find_data(file, n_init = 25)
}
\arguments{
\item{file}{Path to data file from an electrochemical experiment}

\item{n_init}{Initial number of lines to load (the starting point varies from file to file)}
}
\description{
Finds the starting point of the data table in data files from electrochemical experiments
\code{n_init} is recursively doubled until the starting point is found.
}
