% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrochemistry.R
\name{get_exp_type}
\alias{get_exp_type}
\title{Return electrochemical experiment type}
\usage{
get_exp_type(file)
}
\arguments{
\item{file}{Path to data file from an electrochemical experiment}
}
\description{
Return electrochemical experiment type
}
