\name{landcover}
\alias{landcover}
\docType{data}
\title{
Fictional landcover data to demonstrate the cca for Raster-Data
}
\description{
A fictional landcover dataset with six different landtypes indicated by number from 0 to 5. Cells with value one a considered as urban in the examples.
}
\usage{data(landcover)}
\format{
  The format is:
Formal class 'RasterLayer' [package "raster"] with 12 slots
}
\examples{
data(landcover)
str(landcover)
}
\keyword{datasets}
