% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cv}
\alias{plot.cv}
\title{Plot cyclic voltammograms}
\usage{
\method{plot}{cv}(df)
}
\arguments{
\item{df}{A CV loaded with \code{cv_read()}}
}
\description{
Produces a CV-plot based on a dataframe from \code{\link{cv_read}}.
If there is more than one CV in the file, each will be plotted in a different color.
}
\examples{
file <- system.file('extdata/cv/5cv_example.txt', package = 'osc')
df <- cv_read(file, skip = 70)
plot(df)
}
\seealso{
Other cyclic voltammetry: \code{\link{cv_read}}
}
