% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrochemistry.R
\name{plot.electrolysis}
\alias{plot.electrolysis}
\title{Plot electrolysis}
\usage{
\method{plot}{electrolysis}(df)
}
\arguments{
\item{df}{An electrolysis loaded with \code{electrolysis_read()}}
}
\description{
Produces an electrolysis based on a dataframe from \code{\link{electrolysis_read}}.
}
