% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.raman_curvefit}
\alias{plot.raman_curvefit}
\title{Plot results of graphene curve fit}
\usage{
\method{plot}{raman_curvefit}(df, ..., col_name)
}
\arguments{
\item{df}{A curvefit dataframe from \code{raman_curvefit_read()}}

\item{col_name}{The name of the column to plot}
}
\description{
Plots a Raman map, if given a raman curvefit from \code{\link{raman_curvefit_read}} and a column.
}
\examples{
df <- raman_curvefit_read(system.file('extdata/graphene_curve_fit_export', package = 'gRaphene'))
plot(df, col_name = `D int`)
}
\seealso{
Other curve fit functions: \code{\link{raman_curvefit_read}}
}
\keyword{curve}
\keyword{fit}
\keyword{raman,}
