% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{plot_area}
\alias{plot_area}
\title{Plot fitted area}
\usage{
plot_area(df)
}
\description{
Plot fitted area
}
\examples{
file <- system.file("extdata/cv/cv_example.txt", package = "osc")
df <- echem_read(file)

df1 <- area(df, sw = 1, x1 = -1.85, x2 = -1.40)
g1 <- plot_area(df1)
df2 <- area(df, sw = 2, x1 = -1.65, x2 = -1.25)
g2 <- plot_area(df2)

gridExtra::grid.arrange(g1, g2, ncol = 1)
}
\seealso{
Other area functions: \code{\link{area}}
}
