\name{population}
\alias{population}
\docType{data}
\title{
Example population data for the  city clustering algorithm 
}
\description{
Example population data for the  city clustering algorithm 
}
\usage{data(population)}
\format{
  The format is:
 num [1:1525, 1:1000] 0 0 0 0 0 0 0 0 0 0 ...
}
\examples{
data(population)
str(population)
}
\keyword{datasets}
