% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{area}
\alias{area}
\title{Fit a background and calculate the area between to points}
\usage{
area(df, sw, x1, x2, p = 3, span = 0.01)
}
\description{
Fit a background and calculate the area between to points
}
