% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cv_read}
\alias{cv_read}
\title{Load cyclic voltammogram}
\usage{
cv_read(file, skip, col_names = c("pot", "cur"), ...)
}
\arguments{
\item{skip}{Number of lines to skip (where does metadata end and the data start?)}

\item{col_names}{A vector containing the names to be used for the two loaded columms (Default: c("pot","cur))}

\item{path}{Path to the CV file (txt)}
}
\description{
Reads *.txt files from the CHI potentiostats. Automatically detects sweep- and cv-number, so these can later be used for colors in plots etc.
}
\examples{
file <- system.file('extdata/cv/cv_example.txt', package = 'osc')
df <- cv_read(file, skip = 41)

file <- system.file('extdata/cv/5cv_example.txt', package = 'osc')
df <- cv_read(file, skip = 70)
plot(df)
}
\seealso{
Other cyclic voltammetry: \code{\link{plot.cv}}
}
\keyword{cyclic}
\keyword{electrochemistry}
\keyword{voltammetry,}
