% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.raman_curvefit}
\alias{plot.raman_curvefit}
\title{Plot results of graphene curve fit}
\usage{
\method{plot}{raman_curvefit}(df, ..., col_name)
}
\description{
Plot results of graphene curve fit
}
\examples{
raman_curvefit_read("data")
raman_curvefit_read(system.file('extdata/graphene_curve_fit_export', package = 'gRaphene'))
}
\seealso{
Other curve fit functions: \code{\link{raman_curvefit_read}}
}
\keyword{curve}
\keyword{fit}
\keyword{raman,}
